Code refactoring - extract ResolutionConfig
