+++ uploader/src/main/scala/com/gu/media/upload/mediaconvert/file/FileOutputGroup.scala
@@ -1,4 +1,5 @@
 package com.gu.media.upload.mediaconvert.file
+import com.gu.media.upload.mediaconvert.Resolution
 import com.gu.media.upload.mediaconvert.OutputGroupDefinition
 import software.amazon.awssdk.services.mediaconvert.model.{
   FileGroupSettings,
