+++ uploader/src/main/scala/com/gu/media/upload/mediaconvert/file/MP4Output.scala
@@ -8,24 +8,7 @@ import com.gu.media.upload.mediaconvert.SharedCodecSettings.{
   h264Settings
 }
 import software.amazon.awssdk.services.mediaconvert.model._
-case class Dimensions(width: Option[Int], height: Option[Int])
-
-sealed trait ResolutionConfig {
-  def dimensions: Dimensions
-  def nameModifier: String
-}
-
-object Resolution {
-  case object High extends ResolutionConfig {
-    val dimensions = Dimensions(None, Some(720))
-    val nameModifier = "_720h"
-  }
-
-  case object Low extends ResolutionConfig {
-    val dimensions = Dimensions(Some(480), None)
-    val nameModifier = "_480w"
-  }
-}
+import com.gu.media.upload.mediaconvert.ResolutionConfig
 
 object MP4Output {
